chore: improve landing page maintainability
