+++ components/FiveOclockDeco/HeroSection.tsx
@@ -53,6 +53,7 @@ export function HeroSection({ location, digitRefs, timerRef }: HeroSectionProps)
           strokeLinecap="round"
           strokeLinejoin="round"
           aria-hidden="true"
+          focusable="false"
         >
           <polyline points="6 9 12 15 18 9" />
         </svg>
