+++ components/FiveOclockDeco/game-sections.css
@@ -1,0 +1,169 @@
+/* --- Section Dividers --- */
+.section-divider {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 20px;
+  margin: 60px 0 40px;
+}
+.section-divider::before,
+.section-divider::after {
+  content: "";
+  height: 1px;
+  width: 100%;
+  max-width: 300px;
+  background: linear-gradient(to right, transparent, var(--gold-text), transparent);
+  opacity: 0.6;
+}
+.section-divider h2 {
+  font-size: 2rem;
+  letter-spacing: 4px;
+  white-space: nowrap;
+  font-weight: 600;
+}
+
+/* --- Game Night Section --- */
+.game-night {
+  padding-bottom: 60px;
+  text-align: center;
+}
+.game-subtitle {
+  font-size: 1.2rem;
+  color: var(--cream);
+  margin-bottom: 40px;
+}
+.cards-grid {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 30px;
+  margin-bottom: 30px;
+}
+
+.game-card {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+  padding: 2.5rem 1.5rem 2rem;
+  height: 100%;
+  transition:
+    transform 0.5s var(--ease-premium),
+    box-shadow 0.5s ease;
+  will-change: transform;
+}
+/* Fallback hover for touch/no-js */
+@media (hover: hover) {
+  .game-card:hover {
+    transform: translateY(-10px);
+    box-shadow: 0 25px 50px rgba(0, 0, 0, 0.8);
+  }
+}
+
+.game-card::after {
+  content: "";
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  border: 2px solid var(--gold);
+  opacity: 0;
+  transform: scale(0.95);
+  transition: all 0.4s var(--ease-premium);
+  pointer-events: none;
+}
+.game-card:hover::after {
+  opacity: 1;
+  transform: scale(1);
+}
+
+.card-img-container {
+  width: 100%;
+  height: 180px;
+  margin-bottom: 20px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  position: relative;
+  overflow: hidden;
+}
+.card-img-container img {
+  max-width: 100%;
+  max-height: 100%;
+  object-fit: contain;
+  filter: drop-shadow(0 10px 15px rgba(0, 0, 0, 0.5));
+  transition: transform 0.4s ease;
+}
+.game-card:hover .card-img-container img {
+  transform: scale(1.05);
+}
+
+.card-img-container::before {
+  content: "";
+  position: absolute;
+  top: -50%;
+  left: -50%;
+  width: 200%;
+  height: 200%;
+  background: linear-gradient(
+    to right,
+    rgba(255, 255, 255, 0) 0%,
+    rgba(255, 255, 255, 0.15) 50%,
+    rgba(255, 255, 255, 0) 100%
+  );
+  transform: rotate(30deg) translateX(-100%);
+  transition: transform 0.6s ease;
+  pointer-events: none;
+  z-index: 2;
+}
+.game-card:hover .card-img-container::before {
+  transform: rotate(30deg) translateX(100%);
+}
+
+.game-card h3 {
+  font-size: 1.5rem;
+  color: var(--cream);
+  margin-bottom: 10px;
+  letter-spacing: 1px;
+}
+.game-card p {
+  font-size: 1.1rem;
+  color: rgba(220, 205, 169, 0.9);
+  margin-bottom: 20px;
+  flex-grow: 1;
+}
+.no-account-note {
+  font-size: 1.1rem;
+  color: var(--gold-text);
+  font-style: italic;
+  margin-top: 20px;
+  text-shadow: 0 1px 2px rgba(0, 0, 0, 0.8);
+}
+
+@media (max-width: 820px) {
+  .cards-grid {
+    grid-template-columns: 1fr;
+    max-width: 400px;
+    margin-left: auto;
+    margin-right: auto;
+  }
+}
+
+/* --- About Section --- */
+.about {
+  text-align: center;
+  max-width: 800px;
+  margin: 0 auto 60px;
+  padding: 0 20px;
+}
+.about h2 {
+  font-size: 1.5rem;
+  margin-bottom: 20px;
+  letter-spacing: 2px;
+  font-weight: 600;
+}
+.about p {
+  font-size: 1.1rem;
+  color: var(--cream);
+  line-height: 1.8;
+}
