+++ components/FiveOclockDeco/header.css
@@ -1,0 +1,85 @@
+/* --- Header --- */
+header {
+  padding: 30px 0;
+  border-bottom: 1px solid rgba(168, 147, 104, 0.2);
+  margin-bottom: 40px;
+}
+.header-content {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.nav-link {
+  color: var(--gold-text);
+  text-decoration: none;
+  font-size: 0.9rem;
+  letter-spacing: 2px;
+  transition: color 0.3s;
+  position: relative;
+  padding-bottom: 4px;
+}
+.nav-link::after {
+  content: "";
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+  height: 1px;
+  background-color: var(--gold-text);
+  transform: scaleX(0);
+  transform-origin: right;
+  transition: transform 0.3s ease;
+}
+.nav-link:hover::after {
+  transform: scaleX(1);
+  transform-origin: left;
+}
+
+h1.logo-area {
+  text-align: center;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin: 0;
+}
+.logo-top {
+  display: flex;
+  align-items: center;
+  gap: 15px;
+  font-size: 1.2rem;
+  letter-spacing: 3px;
+  font-weight: 400;
+}
+.logo-top::before,
+.logo-top::after {
+  content: "";
+  display: block;
+  width: 40px;
+  height: 1px;
+  background-color: var(--gold-text);
+  opacity: 0.8;
+}
+.logo-bottom {
+  font-size: 1.8rem;
+  letter-spacing: 4px;
+  margin-top: 5px;
+  font-weight: 600;
+}
+
+@media (max-width: 768px) {
+  .header-content {
+    flex-direction: column;
+    gap: 20px;
+  }
+  .logo-top::before,
+  .logo-top::after {
+    width: 20px;
+  }
+  .logo-top {
+    font-size: 1rem;
+  }
+  .logo-bottom {
+    font-size: 1.4rem;
+  }
+}
